Add wire serialization with protobufs
My intention is to create a client that will monitor the state of the
force sensor.  That means I need to send a live signal to a remote
client.  Here's the proposed architecture: I send UDP messages within
the EtherCAT loop to the client.  The messages get serialized with
Protocol Buffers, and then the client renders out something nice with
ImGUI.

Having UDP in the EtherCAT loop is deeply scuffed, but the bandwidth and
network complexity are quite low, so I anticipate that it won't be a
problem.  If it is, I can send the data via a shm transport to a
different process and have it send UDP messages.

I'm choosing UDP because it's light enough that this might just work.