Remove rust-native data abstraction
The protobufs are providing a data abstraction for the ClipX, so there's
no need to do EtherCAT wire -> Rust -> Protobuf when I can just do
EtherCAT wire -> Protobuf.  So this strips out that middle
representation.  This also benefits because Protobufs ensures that I
have a consistent data representation between this server and the
client.

This has the disadvantage that Protobufs doesn't model Rust perfectly.
Some fields are unnecessarily an `Option<T>`, I suppose to support JS et
al.  `Result<T, E>` is also something that I try to use, but it doesn't
produce something like a `std::result::Result`.  Perhaps when
`std::ops::Try` is stabilized, this will be possible.