Send monitoring data over UDP
The goal with monitoring is that you shouldn't have to log in to a
remote shell to see what's happening with the device.  Here I'm sending
the data over UDP in the same loop as the EtherCAT.  This is a bit
janky, but I reckon it should work for this simple case.

I elected to make the client in Python because it doesn't require much
installation on monitoring machines and is relatively familiar to R&D
engineers, who are realistically my users.