Alter the PDO config to match the sensor type
I had been rolling with the box defaults, but Timon tells me that it's
important that the sensor be marked as the right type and that we get
the Electrical signal.  This does involve altering both the Pre-Op
config and the runtime code, so that's an interesting thing that I may
consider in the future.