Improve data abstraction and output CSV
This moves the definitions for the ClipX-related data from Rust into
Protobufs.  This will later allow me to send data from the target
device, the Raspberry Pi, to a remote monitor with the knowledge that
the monitor will be able to parse the data using 3rd party tools.  The
principal issue that came up with this is that I made the data structure
fairly Rusty by using a `oneof` with variants `Ok` and `Err`.  This is a
whole lot like a Rust `Result`, but Prost, the Rust protobufs library,
doesn't parse the structure as a `std::result::Result`.

This also outputs live data to stdout as a CSV.  The implementation is
quite hacky and would be better served with a serde-based method, but
that task is left for a later date.